Clean up Platform on PlatformRenderer disposal: fixes #4671
